functional-tests/grpc-calls/requirements.txt
grpcio-tools==1.60
grpcio==1.60
pytest==7.2.1
